refactor: 중복되는 로직 useSetTargetMessage custom hook 으로 분리후 적용
